+++ lib/veri/controllers/concerns/authentication.rb
@@ -1,5 +1,3 @@
-require "zlib"
-
 module Veri
   module Authentication
     extend ActiveSupport::Concern
@@ -29,9 +27,9 @@ module Veri
     end
 
     def current_session
-      token = cookies.encrypted["#{auth_cookie_prefix}_token"]
+      token = cookies.encrypted["veri_token"]
 
-      @current_session ||= Session.lookup(token, resolved_tenant)
+      @current_session ||= Session.find_active(token, resolved_tenant)
     end
 
     def log_in(authenticatable)
@@ -44,13 +42,15 @@ module Veri
 
       token = Veri::Session.establish(processed_authenticatable, request, resolved_tenant)
 
-      cookies.encrypted.permanent["#{auth_cookie_prefix}_token"] = { value: token, httponly: true }
+      cookies.encrypted.permanent["veri_token"] = { value: token, httponly: true, secure: request.ssl? }
+      reset_memoization
       true
     end
 
     def log_out
       current_session&.terminate
-      cookies.delete("#{auth_cookie_prefix}_token")
+      cookies.delete("veri_token")
+      reset_memoization
     end
 
     def logged_in?
@@ -58,7 +58,7 @@ module Veri
     end
 
     def return_path
-      cookies.signed["#{auth_cookie_prefix}_return_path"]
+      cookies.signed["veri_return_path"]
     end
 
     def shapeshifter?
@@ -68,20 +68,14 @@ module Veri
     private
 
     def with_authentication
-      if logged_in? && current_session.active?
-        if current_user.locked?
-          log_out
-          when_unauthenticated
-        else
-          current_session.update_info(request)
-        end
-
+      if logged_in? && !current_user.locked?
+        current_session.update_info(request)
         return
       end
 
       log_out
 
-      cookies.signed["#{auth_cookie_prefix}_return_path"] = { value: request.fullpath, expires: 15.minutes.from_now } if request.get? && request.format.html?
+      cookies.signed["veri_return_path"] = { value: request.fullpath, expires: 15.minutes.from_now } if request.get? && request.format.html?
 
       when_unauthenticated
     end
@@ -100,8 +94,8 @@ module Veri
       ).resolve
     end
 
-    def auth_cookie_prefix
-      @auth_cookie_prefix ||= "auth_#{Zlib.crc32(Marshal.dump(resolved_tenant))}"
+    def reset_memoization
+      @current_user = @current_session = nil
     end
   end
 end