+++ lib/veri/password/scrypt.rb
@@ -12,6 +12,10 @@ module Veri
       def verify(password, hashed_password)
         ::SCrypt::Password.new(hashed_password) == password
       end
+
+      def match?(hashed_password)
+        hashed_password.match?(/\A\h+\$\h+\$\h+\$/)
+      end
     end
   end
 end